# weather=practice=api-docs.json
{"openapi":"3.0.1","info":{"title":"게시판 API","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"User","description":"User API"}],"paths":{"/api/users":{"post":{"tags":["User"],"summary":"User 등록","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"201":{"description":"User 생성 성공!","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/weather":{"get":{"tags":["weather-controller"],"operationId":"getAllWeather","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WeatherDto"}}}}}}}},"/api/weather/{districtCode}":{"get":{"tags":["weather-controller"],"operationId":"getWeather","parameters":[{"name":"districtCode","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeatherDto"}}}}}}},"/api/users/{userId}":{"get":{"tags":["User"],"summary":"User 조회","operationId":"getUser","parameters":[{"name":"userId","in":"path","description":"UserId","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"User 조회 성공!","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"404":{"description":"User를 찾을 수 없음","content":{"*/*":{"example":"User not found: {userId}"}}}}}}},"components":{"schemas":{"UserRequest":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}},"description":"User 생성 정보"},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"}}},"WeatherDto":{"type":"object","properties":{"district":{"type":"string"},"maxIndex":{"type":"integer","format":"int32"},"airQualityGrade":{"type":"string"},"pollutant":{"type":"string"},"nitrogen":{"type":"number","format":"double"},"ozone":{"type":"number","format":"double"},"carbon":{"type":"number","format":"double"},"sulfurous":{"type":"number","format":"double"},"pm10":{"type":"integer","format":"int32"},"pm25":{"type":"integer","format":"int32"}}}}}}